inference/task_2/requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
xgboost>=1.7.3
scikit-learn>=1.2.2

# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0

# API and server
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pydantic>=1.10.7

# Caching and utilities
redis>=4.5.4
joblib>=1.2.0

# Optional but recommended for better performance
httpx>=0.24.0
aiofiles>=23.1.0

# Development and debugging (optional)
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
requests>=2.28.2
pytorch-lightning==2.4.0